// server/src/models/chatMessage.ts
import { ROOMS } from '../constants/socketEvent';

export type Sender = 'all' | 'alive' | 'spectators';

export interface MessageProps {
    color?: string;
    hideAuthor?: boolean;
}
export interface ChatMessage {
    author: string; // player username or 'Server'
    content: string;
    to?: ROOMS; // undefined for all
    props?: MessageProps;
}

export interface Whisper {
    author: string;
    content: string;
    to: string;
    props?: MessageProps;
}